+++ bass.html
@@ -13,9 +13,7 @@
 
     <script src="socket.io/socket.io.js"></script>
 
-    <script src="src/js/pubsub.js"></script>
-    <script src="src/js/motionTracker.js"></script>
-    <script src="src/js/instruments/bass.js"></script>
+    <script src="dist/bass.js"></script>
   </head>
   <body>
     <div class="frame">
