--- drum-kick.html
@@ -1,28 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Thereminator</title>
-
-    <link rel="stylesheet" type="text/css" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css" href="../src/styles.css">
-
-    <script src="node_modules/tracking/build/tracking-min.js"></script>
-    <script src="node_modules/p5/lib/p5.min.js"></script>
-    <script src="node_modules/p5/lib/addons/p5.sound.js"></script>
-
-    <script src="socket.io/socket.io.js"></script>
-
-    <script src="src/js/pubsub.js"></script>
-    <script src="src/js/motionTracker.js"></script>
-    <script src="src/js/instruments/drum-kick.js"></script>
-  </head>
-  <body>
-    <div class="frame">
-      <div class="container">
-        <video id="video" width="600" height="450" preload autoplay loop muted controls></video>
-        <canvas id="canvas" width="600" height="450"></canvas>
-      </div>
-    </div>
-  </body>
-</html>
