+++ index.html
@@ -9,6 +9,7 @@
       <li><a href="server.html">server</a></li>
       <li><a href="mono.html">mono</a></li>
       <li><a href="oscillator.html">oscillator</a></li>
+      <li><a href="new_bass.html">NEW BASS</a></li>
     </ul>
   </body>
 </html>
