--- src/js/pubsub.js
@@ -1,10 +1,0 @@
-window.pubSub = {
-  listener: null,
-  subscribe: function(fn) {
-    this.listener = fn;
-  },
-
-  init: function(fn) {
-    fn(this.listener);
-  }
-};
